+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.6] - 2025-04-11
+
+### Added
+
+- Adds support for the "<blockquote>" element using the "Goatee.Blockquote()" class.
+- Adds support for the "<code>" element using the "Goatee.Code()" class.
+
 ## [0.3.5] - 2025-04-11
 
 ### Added